fix departing unit subtleties (#29)
ops 1.5.2 introduced a fix to the harness' handling for departing unit
tracking for events.  This uncovered a subtle error in the tests here.
remove_relation was triggering the "departing" relation data key to be
inserted into the local charm's data because the local charm had been
added to the relation units list (which is not really how that is
intended to be used).  This was causing the delete_user method to be
called earlier than intended and caused assertion failures in the
on_relation_broken test.  The fix is to not add the local unit to the
relation units.  But then the relation departed test that checks for the
"departing" data key wasn't working.  So for now, we can add the local
unit to the relation in that test only.